requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.8.0.76
requests==2.31.0
python-dotenv==1.0.0
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2 